tests/requirements.txt
# 测试依赖

# 性能测试
locust>=24.0.0
psutil>=5.9.0

# 模糊测试
hypothesis>=6.100.0

# 代码覆盖率
pytest-cov>=4.1.0

# 异步测试
pytest-asyncio>=0.23.0

# 测试报告
pytest-html>=4.0.0
pytest-json-report>=1.5.0

# 性能分析
pytest-benchmark>=4.0.0
py-spy>=0.3.14

# 网络测试工具
pytest-socket>=0.6.0

# 内存检测
filprofiler>=2023.0.0
